remove cosmo promise facades
