+++ apps/web/src/lib/functions/artists.ts
@@ -1,6 +1,7 @@
 import { remember } from "@/lib/server/cache.server";
 import { getProxiedToken } from "@/lib/server/proxied-token.server";
 import { getRequestSignal } from "@/lib/server/request.server";
+import { runCosmo } from "@apollo/cosmo/runtime";
 import { fetchArtist, fetchArtists } from "@apollo/cosmo/server/artists";
 import type { CosmoArtistWithMembersBFF } from "@apollo/cosmo/types/artists";
 import { createServerFn } from "@tanstack/react-start";
@@ -16,9 +17,11 @@ export const $fetchArtists = createServerFn({ method: "GET" }).handler(
     const signal = getRequestSignal();
     return await remember(ARTISTS_CACHE_KEY, 60 * 60, async () => {
       const { accessToken } = await getProxiedToken(signal);
-      const artists = await fetchArtists(accessToken, signal);
+      const artists = await runCosmo(fetchArtists(accessToken), signal);
       const withMembers = await Promise.all(
-        artists.map((artist) => fetchArtist(accessToken, artist.name, signal)),
+        artists.map((artist) =>
+          runCosmo(fetchArtist(accessToken, artist.name), signal),
+        ),
       );
       const artistMap: Record<string, CosmoArtistWithMembersBFF> = {};
       withMembers.forEach((artist) => {
