+++ packages/cosmo/package.json
@@ -7,7 +7,7 @@
   "exports": {
     "./types/*": "./src/types/*.ts",
     "./server/*": "./src/server/*.ts",
-    "./effect/*": "./src/effect/*.ts",
+    "./runtime": "./src/runtime.ts",
     "./errors": "./src/errors.ts",
     "./bands": "./src/bands.ts"
   },
