--- packages/cosmo/src/effect/gravity.ts
@@ -1,52 +1,0 @@
-import { Effect } from "effect";
-import {
-  GravityListSchema,
-  GravityResponseSchema,
-  PollDetailResponseSchema,
-} from "../schema/gravity";
-import type { ValidArtist } from "../types/common";
-import { bearer, cosmoClient, decodeBody } from "./http";
-
-/**
- * Fetch the list of gravities for the given artist.
- */
-export const fetchGravities = Effect.fn("Cosmo.fetchGravities")(function* (
-  token: string,
-  artistId: ValidArtist,
-) {
-  const client = yield* cosmoClient;
-  return yield* client
-    .get("/bff/v3/gravities", {
-      headers: bearer(token),
-      urlParams: { artistId },
-    })
-    .pipe(Effect.flatMap(decodeBody(GravityListSchema)), Effect.scoped);
-});
-
-/**
- * Fetch a single gravity.
- */
-export const fetchGravity = Effect.fn("Cosmo.fetchGravity")(function* (
-  token: string,
-  gravityId: number,
-) {
-  const client = yield* cosmoClient;
-  const response = yield* client
-    .get(`/bff/v3/gravities/${gravityId}`, { headers: bearer(token) })
-    .pipe(Effect.flatMap(decodeBody(GravityResponseSchema)), Effect.scoped);
-  return response.gravity;
-});
-
-/**
- * Fetch the poll fields.
- */
-export const fetchPoll = Effect.fn("Cosmo.fetchPoll")(function* (
-  token: string,
-  pollId: number,
-) {
-  const client = yield* cosmoClient;
-  const response = yield* client
-    .get(`/bff/v3/polls/${pollId}`, { headers: bearer(token) })
-    .pipe(Effect.flatMap(decodeBody(PollDetailResponseSchema)), Effect.scoped);
-  return response.pollDetail;
-});
