+++ packages/cosmo/src/server/artists.ts
@@ -1,24 +1,35 @@
-import * as artists from "../effect/artists";
+import { Effect } from "effect";
+import {
+  CosmoArtistListSchema,
+  CosmoArtistWithMembersBFFSchema,
+} from "../schema/artists";
 import type { ValidArtist } from "../types/common";
-import { runCosmo } from "./runtime";
+import { bearer, cosmoClient, decodeBody } from "./http";
 
 /**
  * Fetch artists within COSMO.
  */
-export async function fetchArtists(
+export const fetchArtists = Effect.fn("Cosmo.fetchArtists")(function* (
   token: string,
-  signal: AbortSignal | null = null,
 ) {
-  return await runCosmo(artists.fetchArtists(token), signal);
-}
+  const client = yield* cosmoClient;
+  return yield* client
+    .get("/bff/v3/artists", { headers: bearer(token) })
+    .pipe(Effect.flatMap(decodeBody(CosmoArtistListSchema)), Effect.scoped);
+});
 
 /**
  * Fetch a single artist and its members.
  */
-export async function fetchArtist(
+export const fetchArtist = Effect.fn("Cosmo.fetchArtist")(function* (
   token: string,
   artistId: ValidArtist,
-  signal: AbortSignal | null = null,
 ) {
-  return await runCosmo(artists.fetchArtist(token, artistId), signal);
-}
+  const client = yield* cosmoClient;
+  return yield* client
+    .get(`/bff/v3/artists/${artistId}`, { headers: bearer(token) })
+    .pipe(
+      Effect.flatMap(decodeBody(CosmoArtistWithMembersBFFSchema)),
+      Effect.scoped,
+    );
+});
