+++ packages/cosmo/src/server/collection.ts
@@ -1,25 +1,35 @@
-import * as collection from "../effect/collection";
+import { Effect } from "effect";
+import { ObjektSummariesResponseSchema } from "../schema/collection";
 import type { ValidArtist } from "../types/common";
-import { runCosmo } from "./runtime";
+import { cosmoClient, decodeBody } from "./http";
 
 type FetchSummariesParams = {
   session: string;
   artistId: ValidArtist;
   className: string;
-  signal?: AbortSignal | null;
 };
 
 /**
  * Fetch objekt summaries for a given artist and class.
  */
-export async function fetchObjektSummaries({
-  session,
-  artistId,
-  className,
-  signal = null,
-}: FetchSummariesParams) {
-  return await runCosmo(
-    collection.fetchObjektSummaries({ session, artistId, className }),
-    signal,
-  );
-}
+export const fetchObjektSummaries = Effect.fn("Cosmo.fetchObjektSummaries")(
+  function* ({ session, artistId, className }: FetchSummariesParams) {
+    const client = yield* cosmoClient;
+    const response = yield* client
+      .get("/bff/v3/objekt-summaries", {
+        headers: { cookie: `user-session=${session}` },
+        urlParams: {
+          artistId,
+          "class[]": className,
+          order: "newest",
+          page: "1",
+          size: "30",
+        },
+      })
+      .pipe(
+        Effect.flatMap(decodeBody(ObjektSummariesResponseSchema)),
+        Effect.scoped,
+      );
+    return response.collections;
+  },
+);
